requirements.txt
numpy
datasets
langchain
openai
faiss-cpu
json